+++ fe/src/pages/AddIssue.jsx
@@ -1,0 +1,5 @@
+const AddIssue = () => {
+  return <div>Test</div>;
+};
+
+export default AddIssue;
